Refactor WebComponentRegistry
- add aliasing for classes to prevent searching for a class every time
